Project(Msg Auto)/requirements.txt
streamlit==1.28.1
twilio==8.10.0
pandas==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0 